Simplify introduction a bit more
Drops the paragraph on long-term antigenic evolution (which is outside
the scope of this paper) in favor of a couple sentences leading into our
approaches to phylogenetic visualizations of antigenic evolution. Uses
the references from the original paragraph (Felsenstein, Lemey), adds a
reference to Smith et al. 2004, and adds a reference to Steinbruck et
al. as part of recent work in this area.