Order grouping labels by reference clade and date
Adds a custom script and workflow rule that creates a measurements
configuration JSON with predefined orderings for each grouping column
such that grouping values are sorted by the clade of the reference
strain in descending order of the earliest reference strain. This logic
effectively orders reference strains (and all other grouping labels) by
the order that clades appear in the tree such that the latest clades
appear at the top of the measurements panel. Within each clade, grouping
labels are ordered by the reference strain collection date in descending
order so the latest references in the latest clade appear at the top.