Clarify descriptions of grouping label order
Adds a description of how we order the grouping labels in the analysis
for the paper with a configuration file and rewords the description of
ordering in Results and Figure 2 caption to refer to the measurements
export command's configuration file instead of the sidecar JSON output
itself. This change minimizes the reference to "sidecar JSON" which is a
bit of Nextstrain jargon.