Clarify the contents of the JSON schema
Rewords the reference to the measurements panel JSON schema to clarify
that the document describes the data structure expected by the
measurements panel, since not all readers may be familiar with the term
"JSON schema". Uses a permalink to the schema JSON to increase the
chances that the link will still be valid in the future (for example, we
might eventually rename the `master` branch of augur to `main` and the
original link would break).