Fix bug in data prep guide's GISAID download
Add note to check the box to convert spaces to underscores in the GISAID
sequence download step to avoid a bug in the data prep guide. Without
checking this box, users download sequences with spaces in some record
ids, augur parse silently removes the spaces from the sequence output
and not from the metadata output, and this mismatch causes sequence and
metadata records to become unlinked for all downstream steps.